insightsol/requirements.txt
annotated-types==0.7.0
argcomplete==3.5.2
asgiref==3.8.1
attrs==24.3.0
boto3==1.35.90
botocore==1.35.90
certifi==2024.12.14
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
Django==4.2
django-extensions==3.2.3
django-loguru==1.0.3
djangorestframework==3.15.2
drf-extra-fields==3.7.0
drf-spectacular==0.28.0
durationpy==0.9
filetype==1.2.0
gunicorn==23.0.0
hjson==3.1.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kappa==0.6.0
loguru==0.7.3
MarkupSafe==3.0.2
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.2
placebo==0.9.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
ruff==0.6.9
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.8.3
typing_extensions==4.12.2
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
